refactor: abstract frontend url methods
